refactor: 重构了角色管理的代码，将复杂组件拆分为多个简单组件。增加了角色管理的表格自动刷新的功能，增加了角色管理和后台账号管理的关键操作（增改查）反馈信息提示，
