+++ src/views/PageContainer/UserManage/UserManageAccount/components/UserTable.js
@@ -2,7 +2,7 @@ import ActivationStatusSwitch from "./ActivationStatusSwitch";
 import UserModal from "./UserModal";
 import api from "../../../../../api/user";
 import React from "react";
-import {Table} from "antd";
+import {message, Table} from "antd";
 
 /**
  * 后台账号管理的表格
@@ -19,8 +19,10 @@ export default function UserTable(props) {
         api.UpdateUser(data).then(response => {
             // log 服务端返回的搜索结果
             console.log('updateUserAndRefresh =', response.data)
+            message.success('修改用户信息成功');
         }).catch(error => {
             console.log("error = ",error)
+            message.error('修改用户信息发生错误');
         });
 
         // 刷新表格内容
